# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TON Site Builder API",
    "description": "Backend Telegram Mini App для создания и публикации сайтов в сети TON",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sites": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "List Sites",
        "operationId": "list_sites_api_sites_get",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteListItem"
                  },
                  "title": "Response List Sites Api Sites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Create Site",
        "operationId": "create_site_api_sites_post",
        "parameters": [
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get Site",
        "operationId": "get_site_api_sites__site_id__get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sites"
        ],
        "summary": "Update Site",
        "operationId": "update_site_api_sites__site_id__patch",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sites"
        ],
        "summary": "Delete Site",
        "operationId": "delete_site_api_sites__site_id__delete",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/preview": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Preview Site",
        "description": "HTML-превью ровно тем же рендером, что и публикация — без заливки в Storage.",
        "operationId": "preview_site_api_sites__site_id__preview_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/publish": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Publish",
        "operationId": "publish_api_sites__site_id__publish_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/publish-status": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Publish Status",
        "operationId": "publish_status_api_sites__site_id__publish_status_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/dns-bind": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Dns Bind",
        "description": "Транзакция привязки опубликованного bag id к домену (подписывает владелец).",
        "operationId": "dns_bind_api_sites__site_id__dns_bind_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsBindResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/custom-code/purchase": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Purchase Custom Code",
        "description": "Оплата премиум-блока: цена берётся из активного тарифа kind=custom_code (админка).",
        "operationId": "purchase_custom_code_api_sites__site_id__custom_code_purchase_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/custom-code/confirm": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Confirm Custom Code",
        "description": "Подтверждение оплаты премиум-блока — только после проверки транзакции on-chain.",
        "operationId": "confirm_custom_code_api_sites__site_id__custom_code_confirm_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{site_id}/custom-code": {
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Set Custom Code",
        "description": "Загрузка HTML/CSS/JS. Доступна только после подтверждённой оплаты.",
        "operationId": "set_custom_code_api_sites__site_id__custom_code_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          },
          {
            "name": "X-Telegram-Init-Data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmPaymentRequest": {
        "properties": {
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          },
          "tx_hash": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Tx Hash"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "tx_hash"
        ],
        "title": "ConfirmPaymentRequest"
      },
...
      "DnsBindResponse": {
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TonConnectTransaction"
          }
        },
        "type": "object",
        "required": [
          "transaction"
        ],
        "title": "DnsBindResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewResponse": {
        "properties": {
          "preview_html": {
            "type": "string",
            "title": "Preview Html"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          }
        },
        "type": "object",
        "required": [
          "preview_html"
        ],
        "title": "PreviewResponse"
      },
      "PublishResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "job_id"
        ],
        "title": "PublishResponse"
      },
      "PublishStatusResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SiteStatus"
          },
          "storage_bag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Bag Id"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PublishStatusResponse"
      },
...
      "PurchaseResponse": {
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TonConnectTransaction"
          },
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          }
        },
        "type": "object",
        "required": [
          "transaction",
          "payment_id"
        ],
        "title": "PurchaseResponse"
      },
...
      "SiteCreateRequest": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/SiteType"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "content_json": {
            "type": "object",
            "title": "Content Json"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "title": "SiteCreateRequest"
      },
      "SiteListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/SiteType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "status": {
            "$ref": "#/components/schemas/SiteStatus"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SiteListItem"
      },
...
      "SiteResponse": {
        "properties": {
          "site": {
            "$ref": "#/components/schemas/SiteOut"
          }
        },
        "type": "object",
        "required": [
          "site"
        ],
        "title": "SiteResponse"
      },
...
      "SiteType": {
        "type": "string",
        "enum": [
          "visitka",
          "links",
          "landing",
          "portfolio",
          "events",
          "ton_project"
        ],
        "title": "SiteType"
      },
      "SiteUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content_json": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Json"
          },
          "custom_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomCodeRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "SiteUpdateRequest"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
...
    }
  }
}